pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify-cleaner"
version = "0.1.0"
authors = [
  { name="Bukola", email="bukolaebikefe@gmail.com" },
]
description = "A CLI tool to bulk-unfollow Spotify playlists."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]
dependencies = [
    "click",
    "rich",
    "requests",
    "keyring",
    "python-dotenv",
    "questionary",
]

[project.urls]
"Homepage" = "https://github.com/Tokonigeorge/spotify-cleaner"
"Bug Tracker" = "https://github.com/Tokonigeorge/spotify-cleaner/issues"

[project.scripts]
spotify-cleaner = "spotify_cleaner.cli:main" 